backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
qiskit==1.3.1
numpy==2.2.1
cryptography==44.0.0
pyyaml==6.0.2
prometheus-client==0.21.1
PyJWT==2.10.1
alembic==1.13.1
